+++ src/test/testData/annotator/AnnotatorFile3.glsl
@@ -1,2 +1,3 @@
-#define A 2
-int a = A;
+#define DUP 2
+#define DUP 2
+int a = DUP;